requirements.txt
Django==5.1.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
django-cors-headers==4.4.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1

